Fixed package and description
