+++ src/nomivore/ID.java
@@ -1,0 +1,25 @@
+package nomivore;
+
+public class ID {
+    public static final int
+        FEATHER = 314,
+        SHAFT = 52,
+        HEADLESS = 53,
+        ARROWTIPBRONZE = 39,
+        ARROWTIPIRON = 40,
+        ARROWTIPSTEEL = 41,
+        ARROWTIPMITHRIL = 42,
+        ARROWTIPADAMANT = 43,
+        ARROWTIPRUNE = 44,
+        ARROWTIPAMETHYST = 21350,
+        ARROWTIPDRAGON = 11237,
+        ARROWBRONZE = 882,
+        ARROWIRON = 884,
+        ARROWSTEEL = 886,
+        ARROWMITHRIL = 888,
+        ARROWADAMANT = 890,
+        ARROWRUNE = 892,
+        ARROWAMETHYST = 21326,
+        ARROWDRAGON = 11212,
+        blah = 1;
+}
