--- src/nomivore/scripts/ID.java
@@ -1,60 +1,0 @@
-package scripts;
-
-public class ID {
-    public static final int
-        FEATHER = 314,
-        SHAFT = 52,
-        HEADLESS = 53,
-        GEBANK = 10060,
-        PESTLE = 233,
-        GOATHORN = 9735,
-        GOATHORNDUST = 9736,
-        UNICORNHORN = 237,
-        UNICORNHORNDUST = 235,
-        COMPOSTPOTION4 = 6470,
-        COMPOSTPOTION3 = 6472,
-        COMPOSTPOTION2 = 6474,
-        COMPOSTPOTION1 = 6476,
-        VIAL = 229,
-        COMPOST = 6032,
-        BUCKET = 1925,
-        SUPERCOMPOST = 6034,
-        WILLOWLOGS = 1519,
-        MAPLELOGS = 1517,
-        KNIFE = 946,
-        MAPLELONGU = 62,
-        ARROWTIPBRONZE = 39,
-        ARROWTIPIRON = 40,
-        ARROWTIPSTEEL = 41,
-        ARROWTIPMITHRIL = 42,
-        ARROWTIPADAMANT = 43,
-        ARROWTIPRUNE = 44,
-        ARROWTIPAMETHYST = 21350,
-        ARROWTIPDRAGON = 11237,
-        ARROWBRONZE = 882,
-        ARROWIRON = 884,
-        ARROWSTEEL = 886,
-        ARROWMITHRIL = 888,
-        ARROWADAMANT = 890,
-        ARROWRUNE = 892,
-        ARROWAMETHYST = 21326,
-        ARROWDRAGON = 11212,
-        PASTRYDOUGH = 1953,
-        PIEDISH = 2313,
-        PIESHELL = 2315,
-        blah = 1;
-
-    public static final int[]
-        COMPOSTPOTIONS = {6470,6472,6474,6476},
-        LOGS = {1519, 1519};
-
-    // Rune ess mine IDS
-    public static final int
-        AUBURY = 637,
-        PORTAL = 7475,
-        ESSMINE = 7471,
-        OPENEDDOOR = 11778,
-        CLOSEDDOOR = 11780,
-        RUNEESS = 1436,
-        blahh = 11;
-}
